Fix T514528: An empty action mustn't be wrapped by the scope.$apply method
