+++ js/integration/angular/component_registrator.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var $ = require("jquery"),
+    Config = require("../../core/config"),
     registerComponent = require("../../core/component_registrator"),
     Class = require("../../core/class"),
     Locker = require("../../core/utils/locker"),
@@ -387,16 +388,25 @@ var ComponentBuilder = Class.inherit({
 
             return wrappedAction;
         };
+        result.beforeActionExecute = result.onActionCreated;
         result.nestedComponentOptions = function(component) {
             return {
                 templatesRenderAsynchronously: component.option("templatesRenderAsynchronously"),
+                forceApplyBindings: component.option("forceApplyBindings"),
                 modelByElement: component.option("modelByElement"),
                 onActionCreated: component.option("onActionCreated"),
+                beforeActionExecute: component.option("beforeActionExecute"),
                 nestedComponentOptions: component.option("nestedComponentOptions")
             };
         };
 
         result.templatesRenderAsynchronously = true;
+
+        if(Config().wrapActionsBeforeExecute) {
+            result.forceApplyBindings = function() {
+                safeApply(function() {}, scope);
+            };
+        }
 
         result.integrationOptions = {
             createTemplate: function(element) {
@@ -417,6 +427,10 @@ var ComponentBuilder = Class.inherit({
                 if(!skipCallback) {
                     immediateValue = fn();
                     callback(immediateValue);
+                }
+
+                if(Config().wrapActionsBeforeExecute) {
+                    safeApply(function() {}, scope);
                 }
 
                 return disposeWatcher;
