+++ js/ui/popup.js
@@ -672,6 +672,8 @@ var Popup = Overlay.inherit({
     },
 
     _setContentHeight: function() {
+        (this.option("forceApplyBindings") || $.noop)();
+
         if(this._disallowUpdateContentHeight()) {
             return;
         }
@@ -736,6 +738,8 @@ var Popup = Overlay.inherit({
                 left: 0
             });
         } else {
+            (this.option("forceApplyBindings") || $.noop)();
+
             return this.callBase.apply(this, arguments);
         }
     },
